requirements.txt
django
django-allauth
django-crispy-forms
django-multiselectfield
